requirements.txt
torch>=1.7.0
numpy<2.0.0
opencv-python>=4.1.1
requests>=2.23.0
ipython
Pillow>=7.1.2
psutil
thop>=0.1.1
torchvision>=0.8.1
PyYAML>=5.3.1
matplotlib>=3.2.2
pandas>=1.1.4
seaborn>=0.11.0
tqdm>=4.64.0
pytesseract